Fixed fanout module a little
